AS#160071805908634, Handle jwt token in api calls -- Extracted common logic for adding auth token into base class
